Move RacingCar and RacingGame to domain
