+++ src/main/kotlin/step4/domain/GetWinnerList.kt
@@ -1,7 +1,5 @@
 package step4.domain
 
-import step4.data.RacingCar
-
 object GetWinnerList {
     fun getWinnerList(racingCars: List<RacingCar>): List<RacingCar> {
         val maxMove = racingCars.maxOf { it.moveCount }
