+++ src/test/kotlin/step4/GetWinnerListTest.kt
@@ -2,7 +2,7 @@ package step4
 
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.Test
-import step4.data.RacingCar
+import step4.domain.RacingCar
 import step4.domain.GetWinnerList
 
 class GetWinnerListTest {
